@azu 2016-04-02
Before:

```
Performance Run #1:  1894.025506ms
Performance Run #2:  1665.5913ms
Performance Run #3:  1593.939273ms
Performance Run #4:  1658.1641789999999ms
Performance Run #5:  1488.5748330000001ms
Performance budget ok:  1658.1641789999999ms (limit: 4411.764705882353ms)
```

After:

```
Performance Run #1:  1836.873439ms
Performance Run #2:  1506.06601ms
Performance Run #3:  1537.7413099999999ms
Performance Run #4:  1469.9420850000001ms
Performance Run #5:  1540.071309ms
Performance budget ok:  1537.7413099999999ms (limit: 4411.764705882353ms)
```